feat(#40): utils file split to handlers
